init login/register forms

+++ src/components/Welcome/Register.jsx
@@ -1,20 +1,29 @@
 import React, { useState } from "react";
 const axios = require('axios');
 
-export default function Login() {
+export default function Register() {
 
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [passwordConfirmation, setPasswordConfirmation] = useState('');
+  const [name, setName] = useState('');
 
   const handleSubmit = async (e) => {
     e.preventDefault();
 
     try {
+      if (password !== passwordConfirmation) return console.log('passwords do not match');
+      await axios.post('http://localhost:3001/api/users/register', {
+        email,
+        password,
+        name
+      });
       const response = await axios.post('http://localhost:3001/api/auth/login', {
         email,
         password
-      })
+      });
       localStorage.setItem('userToken', response.data.refreshToken);
+      console.log('User created successfully...');
     } catch (err) {
       console.log (err);
     }
@@ -23,6 +32,16 @@ export default function Login() {
   return (
     <div>
       <form onSubmit={handleSubmit}>
+
+      <label>Name</label>
+        <input
+          type="text"
+          placeholder="Enter name..."
+          name="name"
+          value={name}
+          onChange={(e) => {setName(e.target.value)}}
+        />
+
         <label>Email:</label>
         <input
           type="text"
@@ -31,6 +50,7 @@ export default function Login() {
           value={email}
           onChange={(e) => {setEmail(e.target.value)}}
         />
+
         <label>Password:</label>
         <input 
           type="password"
@@ -39,17 +59,23 @@ export default function Login() {
           value={password}
           onChange={(e) => {setPassword(e.target.value)}}
         />
-        <input type="submit" value="Login" />
+
+        <label>Confirm Password:</label>
+        <input 
+          type="password"
+          placeholder="Enter password..."
+          name="password confirmation"
+          value={passwordConfirmation}
+          onChange={(e) => {setPasswordConfirmation(e.target.value)}}
+        />
+        <input type="submit" value="Register" />
       </form>
 
       <br />
       <br />
       <p>Email: {email}</p>
       <p>Password: {password}</p>
-      <br />
-      <br />
-      <br />
-      <br />
+      <p>Password: {passwordConfirmation}</p>
     </div>
   )
 }
